+++ lib/screen/playing_page.dart
@@ -1,14 +1,14 @@
 import 'package:color_tile/component/constant_block.dart';
 import 'package:color_tile/component/movable_block.dart';
 import 'package:color_tile/constants.dart';
-import 'package:color_tile/provider/score_provider.dart';
-import 'package:color_tile/provider/time_provider.dart';
+import 'package:color_tile/controllers/score_provider.dart';
+import 'package:color_tile/controllers/time_provider.dart';
 import 'package:color_tile/screen/result_page.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:quiver/iterables.dart';
-import 'package:color_tile/model/block.dart';
-import 'package:color_tile/provider/block_provider.dart';
+import 'package:color_tile/controllers/block.dart';
+import 'package:color_tile/controllers/block_provider.dart';
 
 class PlayingPage extends HookConsumerWidget {
   const PlayingPage({super.key});
